Decouple IO for the responder
Two new methods were added to `Writer<'a>` to allow manipulating slices without
having to also carry the corresponding buffer around. This allows easier
decoupling of IO by writing responses directly to the writer and then having a
wrapper method to send the response. The wrapper has the same interface so the
current calling code does not have to change.