Remove unused secret callback, Ref: #624
SpdmRequesterDataSignCbType and SpdmResponderDataSignCbType are identical and these
function definition voilates the sigle responsibilty principle.
To implement them need to look for SPDM spec. for example: op_code parameter.
This will increase the difficulty to implement the interface

Current spdmlib have already use the SpdmAsymSign interface for signing.
The SpdmAsymSign only takes two parameter to determine the private key.
Then the implementation for SpdmAsymSign is only need to use (secret_key, data) for sign operation.

Note: There are restrictions to seek secret key. Fix this is not in the current scope.

Signed-off-by: xiaoyuxlu <xiaoyu1.lu@intel.com>