spdmlib: move `append_message_k/f` to `SpdmContext`
Methods `append_message_k/f` are methods of `SpdmSession`. They need to use
data inside `SpdmContext` in some cases such as mutual authentication.

However, they cannot take the mutable reference of `SpdmContext` because the
`SpdmSession` itself is got from the mutable reference of `SpdmContext` which
is a conflict in rust.

Signed-off-by: Jiaqi Gao <jiaqi.gao@intel.com>